edit inventory page form has state
The Edit page stores the state in the formData variable.  The function handleLineQtyChange is curried so it can be called for an array index first, and then passed to the child PaintEditRow.  PaintEditRow can call this function with vale for qty, and because of the currying the child does not need to provide the index.